refactor(v0.78.0): redaction.py core/cli → core/utils 이전 (의존성 정리 A1, 5/5)
새 사이클 A1. v0.52에서 deferred됐던 33개 import-linter ignore_imports
중 첫 2개 정리.

이전: core/cli/redaction.py (34L) → core/utils/redaction.py
- 단순 regex 기반 API key redactor — CLI 의존성 0
- agent/tool_executor가 redaction을 import하던 cross-layer 위반 해소

호출자 업데이트 (3곳):
- core/agent/tool_executor/executor.py:407
- core/cli/bash_tool.py:145 (bash_tool도 A2에서 이전 예정)
- tests/test_redaction.py:5

pyproject.toml ignore_imports 2 entries 제거:
- "core.agent.tool_executor.executor -> core.cli.redaction" (×2 contracts)

33 → 31 ignore_imports 남음.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (329 source files)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- import-linter 4/4 contracts 깨짐 0
- E2E Cowboy Bebop A (68.4) 변동 없음

5 PR 계획:
- A1 (this) redaction → utils ✅
- A2 bash_tool → core/agent/
- A3 project_detect → core/utils/
- A4 session_checkpoint + transcript → core/runtime_state/ (새 패키지)
- A5 startup → core/lifecycle/
